[💄View] NicknameChangeView 파일 생성 및 기존 사용자 구분 변수 추가
